+++ packages/app/src/assets/audio/StoryFactoryAudio/story_factory_congrats_5_es.mp3
