+++ packages/app/src/pages/Intruder/IntruderGame.tsx
@@ -19,11 +19,11 @@ import { useProfile } from "@/contexts/ProfileContext";
 import almohada1 from "@/assets/icons/intruder_almohada_1.svg";
 import empanada from "@/assets/icons/intruder_empanada.svg";
 import cover from "@/assets/icons/card_back.svg";
-import incorrect_card_audio from "@/assets/audio/intruder_incorrect.wav";
-import correct_card_audio from "@/assets/audio/intruder_correct.wav";
-import card_flip_audio from "@/assets/audio/intruder_card_flip.wav";
-import instruction_en_audio from "@/assets/audio/intruder_game_instruction_en.mp3";
-import instruction_es_audio from "@/assets/audio/intruder_game_instruction_es.mp3";
+import incorrect_card_audio from "@/assets/audio/IntruderAudio/intruder_incorrect.wav";
+import correct_card_audio from "@/assets/audio/IntruderAudio/intruder_correct.wav";
+import card_flip_audio from "@/assets/audio/IntruderAudio/intruder_card_flip.wav";
+import instruction_en_audio from "@/assets/audio/IntruderAudio/intruder_game_instruction_en.mp3";
+import instruction_es_audio from "@/assets/audio/IntruderAudio/intruder_game_instruction_es.mp3";
 import "./Intruder.scss";
 import { useParams } from "react-router";
 import { useFirestore, useFirestoreDocData } from "reactfire";
